Move auth exceptions to the WebActionExceptionMapper
