+++ misk/src/main/kotlin/misk/web/exceptions/WebActionExceptionMapper.kt
@@ -8,6 +8,9 @@ import misk.web.toResponseBody
 import okhttp3.Headers
 import okhttp3.Headers.Companion.toHeaders
 import jakarta.inject.Inject
+import misk.exceptions.UnauthenticatedException
+import misk.exceptions.UnauthorizedException
+import java.net.HttpURLConnection
 
 /**
  * Maps [WebActionException]s into the appropriate status code. [WebActionException]s' response
@@ -17,7 +20,11 @@ internal class WebActionExceptionMapper @Inject internal constructor(
   val config: ActionExceptionLogLevelConfig
 ) : ExceptionMapper<WebActionException> {
   override fun toResponse(th: WebActionException): Response<ResponseBody> {
-    return Response(th.responseBody.toResponseBody(), HEADERS, statusCode = th.code)
+    return when (th) {
+      is UnauthenticatedException -> UNAUTHENTICATED_RESPONSE
+      is UnauthorizedException -> UNAUTHORIZED_RESPONSE
+      else -> Response(th.responseBody.toResponseBody(), HEADERS, statusCode = th.code)
+    }
   }
 
   override fun toGrpcResponse(th: WebActionException): GrpcErrorResponse {
@@ -31,5 +38,17 @@ internal class WebActionExceptionMapper @Inject internal constructor(
   private companion object {
     val HEADERS: Headers =
       listOf("Content-Type" to MediaTypes.TEXT_PLAIN_UTF8).toMap().toHeaders()
+
+    val UNAUTHENTICATED_RESPONSE = Response(
+      "unauthenticated".toResponseBody(),
+      HEADERS,
+      HttpURLConnection.HTTP_UNAUTHORIZED
+    )
+
+    val UNAUTHORIZED_RESPONSE = Response(
+      "unauthorized".toResponseBody(),
+      HEADERS,
+      HttpURLConnection.HTTP_FORBIDDEN
+    )
   }
 }
